Polish nat naming/file structure
